기존 Weapon에 있는 Tracer를 Bullet Tracer로 수정

+++ Source/MyProject/MyWeaponData.h
@@ -52,7 +52,7 @@ class MYPROJECT_API UMyWeaponData : public UMyItemData
 	GENERATED_BODY()
 public:
 	UPROPERTY(EditAnywhere)
-	float Damage;
+	uint16 Damage;
 	UPROPERTY(EditAnywhere)
 	float FireRate;
 	UPROPERTY(EditAnywhere)
